Moving content deployment steps to Contents module
