+++ src/OrchardCore.Modules/OrchardCore.Contents/Deployment/ContentDeploymentSource.cs
@@ -2,10 +2,11 @@ using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
 using OrchardCore.ContentManagement;
 using OrchardCore.ContentManagement.Records;
+using OrchardCore.Deployment;
 using YesSql;
 using YesSql.Services;
 
-namespace OrchardCore.Deployment.Steps
+namespace OrchardCore.Contents.Deployment
 {
     public class ContentDeploymentSource : IDeploymentSource
     {
