+++ src/OrchardCore.Modules/OrchardCore.Contents/Deployment/ContentDeploymentStep.cs
@@ -1,4 +1,6 @@
-namespace OrchardCore.Deployment.Steps
+using OrchardCore.Deployment;
+
+namespace OrchardCore.Contents.Deployment
 {
     /// <summary>
     /// Adds all content items of a specific type to a <see cref="DeploymentPlanResult"/>.
