+++ src/OrchardCore.Modules/OrchardCore.Contents/Views/Items/AllContentDeploymentStep.Fields.Thumbnail.cshtml
