+++ src/OrchardCore.Modules/OrchardCore.Contents/Views/_ViewImports.cshtml
@@ -5,6 +5,11 @@
 @using OrchardCore.ContentManagement.Display.ViewModels
 @using OrchardCore.Contents.Workflows.Activities
 @using OrchardCore.Contents.Workflows.ViewModels
+@using OrchardCore.DisplayManagement.Views
+@using OrchardCore.Deployment
+@using OrchardCore.Contents.Deployment
+@using OrchardCore.Contents.ViewModels
+
 @using OrchardCore.Workflows.Helpers
 @inherits OrchardCore.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
